added ability to iframe-launch component support fallback url, deleted unused files
