+++ packages/one-bx-shell-app/src/app/app.module.ts
@@ -16,15 +16,7 @@ import { APP_ROUTES } from './app.routes';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AppShellModule } from '@fundamental-ngx/app-shell';
 import { HttpClientModule } from '@angular/common/http';
-import { LandingComponent } from './landing/landing.component';
-
-// frank
-import { IFrameLauncherComponent } from './components/iframe-launcher/iframe-launcher.component';
-import { PluginLauncherComponent } from './components/plugin-launcher/plugin-launcher.component';
-
-// our
-import { IframeLauncherComponent } from './iframe-launcher/iframe-launcher.component';
-import { IframeLandingComponent } from './iframe-landing/iframe-landing.component';
+import { COMPONENTS } from './components';
 
 
 @NgModule({
@@ -43,15 +35,7 @@ import { IframeLandingComponent } from './iframe-landing/iframe-landing.componen
   ],
   declarations: [
     AppComponent,
-    LandingComponent,
-
-    // our
-    IframeLandingComponent,
-    IframeLauncherComponent,
-
-    // frank
-    IFrameLauncherComponent,
-    PluginLauncherComponent
+    ...COMPONENTS
   ],
   bootstrap: [AppComponent]
 })
