+++ packages/one-bx-shell-app/src/app/app.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
-import { LandingComponent } from './landing/landing.component';
 import { PluginPageLauncherComponent } from '@fundamental-ngx/app-shell';
-import { IframeLandingComponent } from './iframe-landing/iframe-landing.component';
+import { LandingComponent } from './components';
+import { IframeLandingComponent } from './components';
 
 
 export const APP_ROUTES: Routes = [
