+++ packages/one-bx-shell-app/src/app/components/iframe-landing/iframe-landing.component.html
@@ -1,10 +1,10 @@
 <fd-layout-grid [col]="6">
   <fd-layout-panel fdLayoutGridSpan [columnSpan]="4">
     <fd-layout-panel-body>
-      <fds-iframe-launcher2
+      <fds-iframe-launcher
         app="contentReq" component="prlist"
-        [attrs]="{frameborder:0, width: '100%', height: '350px'}"
-      ></fds-iframe-launcher2>
+        [attrs]="{frameborder:0, width: '100%', height: '350px'}">
+      </fds-iframe-launcher>
     </fd-layout-panel-body>
   </fd-layout-panel>
   <fd-layout-panel fdLayoutGridSpan [columnSpan]="2">
@@ -25,19 +25,19 @@
   </fd-layout-panel>
 
   <fd-layout-panel fdLayoutGridSpan [columnSpan]="5">
-    <fds-iframe-launcher2
+    <fds-iframe-launcher
       app="contentItem" component="recommendedCategories"
-      [attrs]="{frameborder:0, width: '100%', height: '450px'}"
-    ></fds-iframe-launcher2>
+      [attrs]="{frameborder:0, width: '100%', height: '450px'}">
+    </fds-iframe-launcher>
 
   </fd-layout-panel>
 
   <fd-layout-panel fdLayoutGridSpan [columnSpan]="1">
     <fd-layout-panel-body>
-      <fds-iframe-launcher2
+      <fds-iframe-launcher
         app="contentItem2" component="yourFavorite"
-        [attrs]="{frameborder:0, width: '258px', height: '450px'}"
-      ></fds-iframe-launcher2>
+        [attrs]="{frameborder:0, width: '258px', height: '450px'}">
+      </fds-iframe-launcher>
     </fd-layout-panel-body>
   </fd-layout-panel>
 </fd-layout-grid>
