--- packages/one-bx-shell-app/src/app/components/iframe-launcher/iframe-launcher.component.html
@@ -1,11 +1,0 @@
-<div class="responsive-wrapper"
-     [style.minHeight]="height"
->
-  <iframe #iframe
-          class="embedded-iframe"
-          [src]="_iFrameUrl"
-          scrolling="no"
-          frameborder="no">
-  </iframe>
-</div>
-
