+++ packages/one-bx-shell-app/src/app/components/iframe-launcher/iframe-launcher.component.scss
@@ -1,15 +1,6 @@
 .responsive-wrapper {
   position: relative;
   overflow: hidden;
-  with: 100%;
-  height: 100%;
-}
-
-.embedded-iframe {
-  position: absolute;
-  top: 0;
-  left: 0;
   width: 100%;
   height: 100%;
-  border: 0;
 }
