+++ packages/one-bx-shell-app/src/app/components/iframe-launcher/iframe-launcher.component.ts
@@ -1,120 +1,119 @@
 import {
-  AfterViewInit,
   ChangeDetectionStrategy,
   Component,
   ElementRef,
   Input,
   OnChanges,
-  OnInit,
   Renderer2,
-  SimpleChanges,
-  ViewChild
+  ViewChild,
+  AfterViewInit
 } from '@angular/core';
-import {DomSanitizer, SafeUrl} from '@angular/platform-browser';
+import { DomSanitizer, SafeUrl } from '@angular/platform-browser';
+import { LookupService } from '@fundamental-ngx/app-shell';
 
+interface MFFrameConfiguration {
+  name: string;
+  module?: string;
+  entry: string;
+  iframe: string;
 
-/**
- * This should be part of the plugin launcher and depending on the framework type we should either instantiate
- * remote all, or use directly this iFrame wrapper.
- *
- * IFrame launcher loads directly legacy application and they are not exposed as Module Federation plugin.
- * If these legacy application would have such capability there would be no need this component;
- *
- *
- *
- * We will extend PluginDescriptor with additional properties and if framework is Legacy we will handle this by iFrame
- * component. All this will be handled by
- *
- * <fdp-plugin-launcher name="fieldGlass.LandingLinks">
- * </fdp-plugin-launcher>
- *
- * This component can be used also as standalone
- *
- * Todo:
- *  - Finish fds-iframe-launcher standalone component, abstract whatever needs to be abstracted
- *  - Expand existing communication to some API
- *    AppShell -> iFrameLauncher -> iFrame contentWindow
- *    iFrame contentWindow -> iFrameLauncher -> the host app
- *   - Make it part of the plugin luncher and extend configuration.
- *
- *
- *
- *
- */
+  scope?: string;
+  category?: string;
+  hasRoutes?: boolean;
+}
+
+interface MFAppConfiguration {
+  id: string;
+  provider?: string;
+  version: string;
+  host: string;
+  components: Record<string, MFFrameConfiguration>;
+}
+
 @Component({
   selector: 'fds-iframe-launcher',
-  templateUrl: './iframe-launcher.component.html',
   styleUrls: ['./iframe-launcher.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  template: `
+    <iframe
+      #iframe
+      class="responsive-wrapper"
+      [style.minHeight]="attrs.height"
+      load="onLoad(iframe)"
+      *ngIf="iconfig?.src"
+      [src]="iconfig.src"
+    ></iframe>
+  `
 })
-export class IFrameLauncherComponent implements OnInit, OnChanges, AfterViewInit {
-  @Input()
-  url: string;
+export class IframeLauncherComponent implements OnChanges, AfterViewInit {
+  @ViewChild('iframe', {static: false}) iframeEl: ElementRef;
+  iconfig: { src: string, attrs: Record<string, string> };
 
-  @Input()
-  id: string;
-
-  @Input()
-  title: string;
-
-  @Input()
-  height: string;
-
-  @Input()
-  attrs: Record<string, string | number>;
-
-  @ViewChild('iframe', {static: false})
-  iframeEl: ElementRef;
-
-  _iFrameUrl: SafeUrl;
+  @Input() url: string;
+  @Input() app: string;
+  @Input() component: string;
+  @Input() attrs: Record<string, string | number>;
 
   private iframeHandle: any;
 
-  // tslint:disable-next-line:variable-name
-  constructor(private _render: Renderer2, private _sanitizer: DomSanitizer) {
+  constructor(
+    private readonly render: Renderer2,
+    private readonly lookupService: LookupService,
+    private readonly sanitizer: DomSanitizer
+  ) {
   }
 
-  ngOnInit(): void {
-  }
-
-
-  ngOnChanges(changes: SimpleChanges): void {
-    if (changes.url) {
-      this._iFrameUrl = this._sanitizer.bypassSecurityTrustResourceUrl(this.url);
+  ngOnChanges(): void {
+    if (this.app && this.component) {
+      const iconfig = this.getIframeConfig();
+      this.updateAttrs(iconfig.attrs, this.iconfig?.attrs);
+      this.iconfig = iconfig;
+    } else if (this.url) {
+      const iframeUrl = this.sanitizer.bypassSecurityTrustResourceUrl(this.url);
+      this.iconfig = {
+        src: iframeUrl as string,
+        attrs: Object.assign({}, this.attrs) as Record<string, string>
+      }
+      this.updateAttrs(this.iconfig.attrs, this.iconfig?.attrs);
     }
-    this.updateAttrs(this.attrs);
   }
 
   ngAfterViewInit(): void {
-    this.updateAttrs(this.attrs);
-
-    this.setupCommunicationChannel();
-
-
-    setTimeout(() => {
-      this.iframeHandle.postMessage('Hello From AppShell', 'http://localhost:5400');
-    }, 5000);
+    this.updateAttrs(this.iconfig.attrs);
   }
 
   private setupCommunicationChannel(): void {
     if (this.iframeEl) {
       this.iframeHandle = this.iframeEl.nativeElement.contentWindow;
-      window.addEventListener('message', this.onMessage, false);
     }
   }
 
-  private onMessage(e: any): void {
-    console.log('This is iFrame Launcher received message from iFrame => ', e);
-  }
-
-  private updateAttrs(newValue: Record<string, string | number>): void {
+  private updateAttrs(newValue: Record<string, string>, oldValue?: Record<string, string>): void {
     if (!this.iframeEl) {
       return;
     }
-    if (newValue) {
-      for (const key of Object.keys(newValue)) {
-        this._render.setAttribute(this.iframeEl.nativeElement, key, `${newValue[key]}`);
+    if (oldValue) {
+      for (const key of Object.keys(oldValue)) {
+        this.render.removeAttribute(this.iframeEl.nativeElement, key);
       }
     }
+    if (newValue) {
+      for (const key of Object.keys(newValue)) {
+        this.render.setAttribute(this.iframeEl.nativeElement, key, `${ newValue[key] }`);
+      }
+    }
+  }
+
+  private getIframeConfig(): any {
+    const appConfig = this.lookupService
+      // todo: valorkin: this configuration seems to be more appropriate
+      .lookup(new Map([['name', this.app]])) as unknown as { descriptor: MFAppConfiguration };
+    const frameConfig = appConfig.descriptor.components[this.component];
+    const iframeUrl = this.sanitizer
+      .bypassSecurityTrustResourceUrl(`${ appConfig.descriptor.host }${ frameConfig.iframe }`);
+    return {
+      src: iframeUrl,
+      attrs: Object.assign({}, this.attrs)
+    };
   }
 }
