+++ packages/one-bx-shell-app/src/app/components/landing/landing.component.html
@@ -6,10 +6,8 @@
   </fd-layout-panel>
   <fd-layout-panel [columnSpan]="2" fdLayoutGridSpan>
     <fd-layout-panel-body>
-      <fds-iframe-launcher id="landing-legacy1"
-                           title="links from external system"
-                           height="210px"
-                           url="http://localhost:5400">
+      <fds-iframe-launcher url="http://localhost:5400"
+                            [attrs]="{frameborder:0, width: '100%', height: '210px'}">
       </fds-iframe-launcher>
     </fd-layout-panel-body>
   </fd-layout-panel>
