+++ packages/one-bx-shell-app/src/app/components/landing/landing.component.scss
